Add helper classes to physics example, update audio handling, update README
